+++ src/main/java/com/example/h2demo/dao/DAO.java
@@ -9,10 +9,10 @@ public interface DAO <T> {
 
     void create(T t);
 
-    Optional<T> get(int id);
+    Optional<T> get(long id);
 
-    void update(T t, int id);
+    void update(T t, long id);
 
-    void delete(int id);
+    void delete(long id);
 
 }
